location double tap fixed
now search by location will start right after user give permissions
