+++ project/feature/main_screen/build.gradle
@@ -55,7 +55,7 @@ dependencies {
     implementation 'androidx.compose.material3:material3'
     implementation 'androidx.compose.foundation:foundation'
     implementation 'androidx.compose.material:material-icons-extended:1.5.0-alpha03'
-    implementation "com.google.accompanist:accompanist-permissions:0.30.0"
+    implementation "com.google.accompanist:accompanist-permissions:0.31.2-alpha"
     //
     //mvvm
     implementation("androidx.activity:activity-ktx:1.7.1")
